refactor: 응답 메시지 직렬화 로직과 클라이언트에 반환 로직 분리
