+++ src/main/java/webserver/handler/RequestHandler.java
@@ -8,7 +8,7 @@ import org.slf4j.LoggerFactory;
 import webserver.httpMessage.htttpRequest.HttpRequest;
 import webserver.httpMessage.httpResponse.HttpResponse;
 import webserver.httpMessage.htttpRequest.RequestFactory;
-import webserver.httpMessage.ResponseTransmitter;
+import webserver.httpMessage.httpResponse.util.ResponseTransmitter;
 import webserver.router.FrontRouter;
 
 public class RequestHandler implements Runnable {
